# docs/swagger.yaml
definitions:
...
  dto.Login:
    properties:
      captcha_code:
        type: string
      captcha_id:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  dto.Pagination:
    properties:
      current:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  echox.Response:
    properties:
      data: {}
      message: {}
    type: object
...
  models.Role:
    properties:
      created_at:
        $ref: '#/definitions/sql.NullTime'
      created_by:
        type: string
      deleted:
        type: boolean
      id:
        type: string
      name:
        type: string
      remark:
        type: string
      role_menus:
        items:
          $ref: '#/definitions/models.RoleMenu'
        type: array
      sequence:
        type: integer
      status:
        maximum: 1
        minimum: -1
        type: integer
      updated_at:
        $ref: '#/definitions/sql.NullTime'
    required:
    - name
    - remark
    - sequence
    - status
    type: object
  models.RoleMenu:
    properties:
      action_id:
        type: string
      created_at:
        $ref: '#/definitions/sql.NullTime'
      deleted:
        type: boolean
      id:
        type: string
      menu_id:
        type: string
      role_id:
        type: string
      updated_at:
        $ref: '#/definitions/sql.NullTime'
    required:
    - action_id
    - menu_id
    - role_id
    type: object
  models.RoleQueryResult:
    properties:
      list:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      pagination:
        $ref: '#/definitions/dto.Pagination'
    type: object
...
  sql.NullTime:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
info:
  contact: {}
paths:
...
  /api/v1/publics/sys/routes:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: bad request
          schema:
            type: string
        "500":
          description: internal error
          schema:
            type: string
      summary: SysRoutes
      tags:
      - Public
  /api/v1/publics/user:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: bad request
          schema:
            type: string
        "500":
          description: internal error
          schema:
            type: string
      summary: UserInfo
      tags:
      - Public
  /api/v1/publics/user/login:
    post:
      parameters:
      - description: Login
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.Login'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: bad request
          schema:
            type: string
        "500":
          description: internal error
          schema:
            type: string
      summary: UserLogin
      tags:
      - Public
...
  /api/v1/publics/user/menutree:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: bad request
          schema:
            type: string
        "500":
          description: internal error
          schema:
            type: string
      summary: UserMenuTree
      tags:
      - Public
  /api/v1/roles:
    get:
      parameters:
      - in: query
        name: current
        type: integer
      - enum:
        - ASC
        - DESC
        in: query
        name: direction
        type: string
        x-enum-varnames:
        - OrderByASC
        - OrderByDESC
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: ids
        type: array
      - in: query
        name: key
        type: string
      - in: query
        name: name
        type: string
      - in: query
        maximum: 128
        name: pageSize
        type: integer
      - in: query
        name: queryValue
        type: string
      - in: query
        maximum: 1
        minimum: -1
        name: status
        type: integer
      - in: query
        name: userID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/echox.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.RoleQueryResult'
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echox.Response'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/echox.Response'
      summary: Role Query
      tags:
      - Role
    post:
      parameters:
      - description: Role
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Role'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/echox.Response'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echox.Response'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/echox.Response'
      summary: Role Create
      tags:
      - Role
  /api/v1/roles.all:
    get:
      parameters:
      - in: query
        name: current
        type: integer
      - enum:
        - ASC
        - DESC
        in: query
        name: direction
        type: string
        x-enum-varnames:
        - OrderByASC
        - OrderByDESC
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: ids
        type: array
      - in: query
        name: key
        type: string
      - in: query
        name: name
        type: string
      - in: query
        maximum: 128
        name: pageSize
        type: integer
      - in: query
        name: queryValue
        type: string
      - in: query
        maximum: 1
        minimum: -1
        name: status
        type: integer
      - in: query
        name: userID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/echox.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Role'
                  type: array
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echox.Response'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/echox.Response'
      summary: Role Get All
      tags:
      - Role
  /api/v1/roles/{id}:
    delete:
      parameters:
      - description: role id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/echox.Response'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echox.Response'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/echox.Response'
      summary: Role Delete By ID
      tags:
      - Role
    get:
      parameters:
      - description: role id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/echox.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Role'
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echox.Response'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/echox.Response'
      summary: Role Get By ID
      tags:
      - Role
    put:
      parameters:
      - description: role id
        in: path
        name: id
        required: true
        type: integer
      - description: Role
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Role'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/echox.Response'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echox.Response'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/echox.Response'
      summary: Role Update By ID
      tags:
      - Role
  /api/v1/roles/{id}/disable:
    put:
      parameters:
      - description: role id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/echox.Response'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echox.Response'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/echox.Response'
      summary: Role Disable By ID
      tags:
      - Role
  /api/v1/roles/{id}/enable:
    put:
      parameters:
      - description: role id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/echox.Response'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echox.Response'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/echox.Response'
      summary: Role Enable By ID
      tags:
      - Role
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"